third commit

+++ Libft/ft_bzero.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(char *s, int n)
+#include "libft.h"
+
+void	ft_bzero(char *s, size_t n)
 {
 	while (n > 0)
 	{
