+++ Libft/ft_memchr.c
@@ -1,7 +1,9 @@
+#include "libft.h"
+
 void	*ft_memchr(const void *ptr, int c, size_t n)
 {
 	unsigned char *str;
-	unsigned int i;
+	size_t i;
 
 	str = (unsigned char *)ptr;
 	c = (unsigned char)c;
