+++ Libft/ft_memcmp.c
@@ -1,8 +1,10 @@
+#include "libft.h"
+
 int	ft_memcmp(const void *ptr1, const void *ptr2, size_t n)
 {
 	char *str1;
 	char *str2;
-	unsigned int i;
+	size_t i;
 
 	str1 = (char*)ptr1;
 	str2 = (char*)ptr2;
